backend/netsutra/requirements.txt
Django>=5.2
djangorestframework>=3.14
django-cors-headers>=4.0
